# apps/server/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/refunds": {
      "get": {
        "operationId": "listRefunds",
        "parameters": [
          {
            "name": "projectId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundListResponseDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/refunds/{id}": {
      "get": {
        "operationId": "getRefund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/customers/{id}/refunds": {
      "post": {
        "operationId": "createRefund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRefundDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/refunds/{id}/actions/approve": {
      "post": {
        "operationId": "approveRefund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveRefundDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/refunds/{id}/actions/process": {
      "post": {
        "operationId": "processRefund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcessRefundDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/refunds/{id}/actions/settle": {
      "post": {
        "operationId": "settleRefund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/refunds/{id}/actions/fail": {
      "post": {
        "operationId": "failRefund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialReversalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/refunds/{id}/actions/cancel": {
      "post": {
        "operationId": "cancelRefund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialReversalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    },
    "/api/v1/refunds/{id}/actions/reverse": {
      "post": {
        "operationId": "reverseRefund",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinancialReversalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundDto"
                }
              }
            }
          }
        },
        "summary": "",
        "tags": ["Finance"]
      }
    }
  },
  "info": {
    "title": "Lincoln Dirt and Gravel API",
    "description": "Lincoln Dirt and Gravel operating system API",
    "version": "1.7.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "FinancialReversalDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 1000
          }
        },
        "required": ["reason"]
      },
...
      "RefundDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "refundNumber": {
            "type": "string"
          },
          "customerAccountId": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "sourceType": {
            "type": "string",
            "enum": ["payment", "deposit", "customer_credit"]
          },
          "sourceId": {
            "type": "string",
            "format": "uuid"
          },
          "amountCents": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "refundMethod": {
            "type": "string"
          },
          "originalMethod": {
            "type": "string",
            "nullable": true
          },
          "payee": {
            "type": "object",
            "additionalProperties": true
          },
          "status": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "alternateMethodReason": {
            "type": "string",
            "nullable": true
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "processedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "settledAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "providerName": {
            "type": "string",
            "nullable": true
          },
          "providerRefundId": {
            "type": "string",
            "nullable": true
          },
          "reversesRefundId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "id",
          "refundNumber",
          "customerAccountId",
          "sourceType",
          "sourceId",
          "amountCents",
          "currency",
          "refundMethod",
          "payee",
          "status",
          "reason"
        ]
      },
      "RefundListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RefundDto"
            }
          }
        },
        "required": ["items"]
      },
      "CreateRefundDto": {
        "type": "object",
        "properties": {
          "sourceType": {
            "type": "string",
            "enum": ["payment", "deposit", "customer_credit"]
          },
          "sourceId": {
            "type": "string",
            "format": "uuid"
          },
          "amountCents": {
            "type": "number",
            "minimum": 1
          },
          "refundMethod": {
            "type": "string",
            "enum": [
              "cash",
              "zelle",
              "venmo",
              "cash_app",
              "paypal",
              "card",
              "bank_transfer",
              "check"
            ]
          },
          "payeeName": {
            "type": "string",
            "maxLength": 200
          },
          "payeeEmail": {
            "type": "string",
            "maxLength": 320
          },
          "reason": {
            "type": "string",
            "maxLength": 1000
          },
          "alternateMethodReason": {
            "type": "string",
            "maxLength": 1000
          }
        },
        "required": ["sourceType", "sourceId", "amountCents", "refundMethod", "payeeName", "reason"]
      },
      "ApproveRefundDto": {
        "type": "object",
        "properties": {
          "identityVerificationReference": {
            "type": "string",
            "maxLength": 300
          }
        }
      },
      "ProcessRefundDto": {
        "type": "object",
        "properties": {
          "providerName": {
            "type": "string",
            "maxLength": 80
          },
          "providerRefundId": {
            "type": "string",
            "maxLength": 200
          }
        }
      }
    }
  }
}